+++ src/main/java/com/github/games647/lagmonitor/LagMonitor.java
@@ -22,20 +22,27 @@ import com.github.games647.lagmonitor.inject.ListenerInjector;
 import com.github.games647.lagmonitor.inject.TaskInjector;
 import com.github.games647.lagmonitor.listeners.PlayerPingListener;
 import com.github.games647.lagmonitor.listeners.ThreadSafetyListener;
+import com.github.games647.lagmonitor.storage.MonitorSaveTask;
+import com.github.games647.lagmonitor.storage.NativeSaveTask;
+import com.github.games647.lagmonitor.storage.Storage;
+import com.github.games647.lagmonitor.storage.TpsSaveTask;
 import com.github.games647.lagmonitor.tasks.BlockingIODetectorTask;
 import com.github.games647.lagmonitor.tasks.PingHistoryTask;
 import com.github.games647.lagmonitor.tasks.TpsHistoryTask;
 import com.github.games647.lagmonitor.traffic.TrafficReader;
 import com.google.common.collect.Maps;
 
+import java.sql.SQLException;
 import java.util.Map;
 import java.util.Timer;
+import java.util.logging.Level;
 
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
 import org.bukkit.plugin.Plugin;
 import org.bukkit.plugin.java.JavaPlugin;
+import org.hyperic.sigar.Sigar;
 
 public class LagMonitor extends JavaPlugin {
 
@@ -51,6 +58,16 @@ public class LagMonitor extends JavaPlugin {
     private TrafficReader trafficReader;
     private Timer blockDetectionTimer;
     private Timer monitorTimer;
+    private Storage storage;
+    private Sigar sigar;
+
+    public LagMonitor() {
+        super();
+
+        //setting the location where sigar can find the library
+        //otherwise it would lookup the library path of Java
+        System.setProperty("org.hyperic.sigar.path", getDataFolder().getPath());
+    }
 
     @Override
     public void onEnable() {
@@ -90,6 +107,34 @@ public class LagMonitor extends JavaPlugin {
             BlockingIODetectorTask blockingIODetectorTask = new BlockingIODetectorTask(this, Thread.currentThread());
             blockDetectionTimer.scheduleAtFixedRate(blockingIODetectorTask, DETECTION_THRESHOLD, DETECTION_THRESHOLD);
         }
+
+        if (sigar != null) {
+            sigar = new Sigar();
+        }
+
+        if (getConfig().getBoolean("monitor-database")) {
+            try {
+                String host = getConfig().getString("host");
+                int port = getConfig().getInt("port");
+                String database = getConfig().getString("database");
+                
+                String username = getConfig().getString("username");
+                String password = getConfig().getString("password");
+                Storage localStorage = new Storage(this, host, port, database, username, password);
+                localStorage.createTables();
+                this.storage = localStorage;
+
+                getServer().getScheduler().runTaskTimer(this, new TpsSaveTask(this), 20L
+                        , getConfig().getInt("tps-save-interval") * 20L);
+                //this can run async because it runs independently from the main thread
+                getServer().getScheduler().runTaskTimerAsynchronously(this, new MonitorSaveTask(this), 20L
+                        , getConfig().getInt("monitor-save-interval") * 20L);
+                getServer().getScheduler().runTaskTimerAsynchronously(this, new NativeSaveTask(this), 20L
+                        , getConfig().getInt("native-save-interval") * 20L);
+            } catch (SQLException sqlEx) {
+                getLogger().log(Level.SEVERE, "Failed to setup monitoring database", sqlEx);
+            }
+        }
     }
 
     @Override
@@ -116,6 +161,10 @@ public class LagMonitor extends JavaPlugin {
         if (securityManager instanceof BlockingSecurityManager) {
             SecurityManager oldSecurityManager = ((BlockingSecurityManager) securityManager).getOldSecurityManager();
             System.setSecurityManager(oldSecurityManager);
+        }
+
+        if (sigar != null) {
+            sigar.close();
         }
 
         for (Plugin plugin : getServer().getPluginManager().getPlugins()) {
@@ -147,6 +196,14 @@ public class LagMonitor extends JavaPlugin {
 
     public PingHistoryTask getPingHistoryTask() {
         return pingHistoryTask;
+    }
+
+    public Sigar getSigar() {
+        return sigar;
+    }
+
+    public Storage getStorage() {
+        return storage;
     }
 
     private void registerCommands() {